Fixed some issues with the Common Mistake component.
